+++ oap-server/server-query-plugin/prometheus-query-plugin/src/main/java/org/apache/skywalking/oap/query/prometheus/PrometheusQueryModule.java
@@ -16,31 +16,19 @@
  *
  */
 
-package org.apache.skywalking.oap.server.core.query.type;
+package org.apache.skywalking.oap.query.prometheus;
 
-import io.vavr.collection.Stream;
-import java.util.ArrayList;
-import java.util.List;
-import lombok.Getter;
+import org.apache.skywalking.oap.server.library.module.ModuleDefine;
 
-public class IntValues {
-    @Getter
-    private List<KVInt> values = new ArrayList<>();
+public class PrometheusQueryModule extends ModuleDefine {
+    public static final String NAME = "query-prometheus";
 
-    public void addKVInt(KVInt e) {
-        values.add(e);
+    public PrometheusQueryModule() {
+        super(NAME);
     }
 
-    public long findValue(String id, int defaultValue) {
-        for (KVInt value : values) {
-            if (value.getId().equals(id)) {
-                return value.getValue();
-            }
-        }
-        return defaultValue;
-    }
-
-    public long latestValue(int defaultValue) {
-        return Stream.ofAll(values).map(KVInt::getValue).findLast(v -> v != defaultValue).getOrElse((long) defaultValue);
+    @Override
+    public Class[] services() {
+        return new Class[0];
     }
 }