+++ oap-server/server-query-plugin/prometheus-query-plugin/src/main/java/org/apache/skywalking/oap/query/prometheus/entity/LabelName.java
@@ -16,31 +16,27 @@
  *
  */
 
-package org.apache.skywalking.oap.server.core.query.type;
+package org.apache.skywalking.oap.query.prometheus.entity;
 
-import io.vavr.collection.Stream;
-import java.util.ArrayList;
-import java.util.List;
-import lombok.Getter;
+public enum LabelName {
+    __name__,
+    layer,
+    scope,
+    service,
+    service_instance,
+    endpoint,
+    //For labeled value query
+    labels,
+    relabels,
+    //For labeled value
+    label,
+    //For tonN and record
+    parent_service,
+    top_n,
+    order,
+    record,
+    //For endpoint_traffic
+    limit,
+    keyword
 
-public class IntValues {
-    @Getter
-    private List<KVInt> values = new ArrayList<>();
-
-    public void addKVInt(KVInt e) {
-        values.add(e);
-    }
-
-    public long findValue(String id, int defaultValue) {
-        for (KVInt value : values) {
-            if (value.getId().equals(id)) {
-                return value.getValue();
-            }
-        }
-        return defaultValue;
-    }
-
-    public long latestValue(int defaultValue) {
-        return Stream.ofAll(values).map(KVInt::getValue).findLast(v -> v != defaultValue).getOrElse((long) defaultValue);
-    }
 }