+++ oap-server/server-query-plugin/prometheus-query-plugin/src/main/java/org/apache/skywalking/oap/query/prometheus/entity/TimeValuePair.java
@@ -16,31 +16,20 @@
  *
  */
 
-package org.apache.skywalking.oap.server.core.query.type;
+package org.apache.skywalking.oap.query.prometheus.entity;
 
-import io.vavr.collection.Stream;
-import java.util.ArrayList;
-import java.util.List;
-import lombok.Getter;
+import com.fasterxml.jackson.databind.annotation.JsonSerialize;
+import lombok.Data;
+import org.apache.skywalking.oap.query.prometheus.entity.codec.TimeValuePairSerializer;
 
-public class IntValues {
-    @Getter
-    private List<KVInt> values = new ArrayList<>();
+@Data
+@JsonSerialize(using = TimeValuePairSerializer.class)
+public class TimeValuePair {
+    private final long time;
+    private String value;
 
-    public void addKVInt(KVInt e) {
-        values.add(e);
-    }
-
-    public long findValue(String id, int defaultValue) {
-        for (KVInt value : values) {
-            if (value.getId().equals(id)) {
-                return value.getValue();
-            }
-        }
-        return defaultValue;
-    }
-
-    public long latestValue(int defaultValue) {
-        return Stream.ofAll(values).map(KVInt::getValue).findLast(v -> v != defaultValue).getOrElse((long) defaultValue);
+    public TimeValuePair(final long time, String value) {
+        this.time = time;
+        this.value = value;
     }
 }