+++ oap-server/server-query-plugin/prometheus-query-plugin/src/main/java/org/apache/skywalking/oap/query/prometheus/entity/response/SeriesQueryRsp.java
@@ -16,31 +16,16 @@
  *
  */
 
-package org.apache.skywalking.oap.server.core.query.type;
+package org.apache.skywalking.oap.query.prometheus.entity.response;
 
-import io.vavr.collection.Stream;
 import java.util.ArrayList;
 import java.util.List;
-import lombok.Getter;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+import org.apache.skywalking.oap.query.prometheus.entity.MetricInfo;
 
-public class IntValues {
-    @Getter
-    private List<KVInt> values = new ArrayList<>();
-
-    public void addKVInt(KVInt e) {
-        values.add(e);
-    }
-
-    public long findValue(String id, int defaultValue) {
-        for (KVInt value : values) {
-            if (value.getId().equals(id)) {
-                return value.getValue();
-            }
-        }
-        return defaultValue;
-    }
-
-    public long latestValue(int defaultValue) {
-        return Stream.ofAll(values).map(KVInt::getValue).findLast(v -> v != defaultValue).getOrElse((long) defaultValue);
-    }
+@EqualsAndHashCode(callSuper = true)
+@Data
+public class SeriesQueryRsp extends QueryResponse {
+    private List<MetricInfo> data = new ArrayList<>();
 }